2nd version of my code
I've amended the code to have the numbers ordered before checking for the gcd. I also changed the output for if one of the numbers is 0 and the other is a positive integer. 
Next I want to make sure that any input that is not positive will still find the gcd.
I also want to prompt the user to add an integer if that was not what they entered the first time.